fix: resolve MyPy typechecking errors in core and mcp server

+++ src/power_framework/core/rot_scoring.py
@@ -56,7 +56,7 @@ def _dense_cosine_similarity(vec_a: list[float], vec_b: list[float]) -> float:
     norm_b = sum(v * v for v in vec_b) ** 0.5
     if norm_a == 0 or norm_b == 0:
         return 0.0
-    return dot / (norm_a * norm_b)
+    return float(dot / (norm_a * norm_b))
 
 
 class ContentDedupDetector:
